Move ValidationErrors to errors package
